Error Logging Update -Correct minor issues, added delete, and documentation
