+++ KFC.Red.Server/KFC.Red.DataAccessLayer/KFC.Red.DataAccessLayer.csproj
@@ -74,6 +74,7 @@
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\Claim.cs" />
+    <Compile Include="Models\ErrorCollection.cs" />
     <Compile Include="Models\MessageDetail.cs" />
     <Compile Include="Models\Question.cs" />
     <Compile Include="Models\Session.cs" />
