+++ KFC.Red.Server/KFC.Red.Dbate.Logging/packages.config
@@ -1,13 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="DnsClient" version="1.0.7" targetFramework="net461" />
-  <package id="MongoDB.Bson" version="2.7.2" targetFramework="net461" />
-  <package id="MongoDB.Driver" version="2.7.2" targetFramework="net461" />
-  <package id="MongoDB.Driver.Core" version="2.7.2" targetFramework="net461" />
+  <package id="DnsClient" version="1.2.0" targetFramework="net472" />
+  <package id="mongocsharpdriver" version="2.8.0" targetFramework="net472" />
+  <package id="MongoDB.Bson" version="2.8.0" targetFramework="net472" />
+  <package id="MongoDB.Driver" version="2.8.0" targetFramework="net472" />
+  <package id="MongoDB.Driver.Core" version="2.8.0" targetFramework="net472" />
   <package id="NLog" version="4.6.0" targetFramework="net461" />
   <package id="NLog.Config" version="4.6.0" targetFramework="net461" />
   <package id="NLog.Mongo" version="4.6.0.68" targetFramework="net461" />
   <package id="NLog.Schema" version="4.6.0" targetFramework="net461" />
-  <package id="System.Buffers" version="4.3.0" targetFramework="net461" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.0.0" targetFramework="net461" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net472" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
 </packages>
